test: move parseURL tests to general_test
