+++ tests/connection_test.ts
@@ -1,4 +1,4 @@
-import { connect, parseURL } from "../redis.ts";
+import { connect } from "../redis.ts";
 import { assertEquals } from "../vendor/https/deno.land/std/testing/asserts.ts";
 import { newClient, startRedis, stopRedis, TestSuite } from "./test_util.ts";
 
@@ -35,60 +35,6 @@ suite.test("select", async () => {
 
 suite.test("swapdb", async () => {
   assertEquals(await client.swapdb(0, 1), "OK");
-});
-
-suite.test("parse basic URL", () => {
-  const options = parseURL("redis://127.0.0.1:7003");
-  assertEquals(options.hostname, "127.0.0.1");
-  assertEquals(options.port, 7003);
-  assertEquals(options.tls, false);
-  assertEquals(options.db, undefined);
-  assertEquals(options.name, undefined);
-  assertEquals(options.password, undefined);
-});
-
-suite.test("parse complex URL", () => {
-  const options = parseURL("rediss://username:password@127.0.0.1:7003/1");
-  assertEquals(options.hostname, "127.0.0.1");
-  assertEquals(options.port, 7003);
-  assertEquals(options.tls, true);
-  assertEquals(options.db, 1);
-  assertEquals(options.name, "username");
-  assertEquals(options.password, "password");
-});
-
-suite.test("parse URL with search options", () => {
-  const options = parseURL(
-    "redis://127.0.0.1:7003/?db=2&password=password&ssl=true",
-  );
-  assertEquals(options.hostname, "127.0.0.1");
-  assertEquals(options.port, 7003);
-  assertEquals(options.tls, true);
-  assertEquals(options.db, 2);
-  assertEquals(options.name, undefined);
-  assertEquals(options.password, "password");
-});
-
-suite.test("Check parameter parsing priority", () => {
-  const options = parseURL(
-    "rediss://username:password@127.0.0.1:7003/1?db=2&password=password2&ssl=false",
-  );
-  assertEquals(options.hostname, "127.0.0.1");
-  assertEquals(options.port, 7003);
-  assertEquals(options.tls, true);
-  assertEquals(options.db, 1);
-  assertEquals(options.name, "username");
-  assertEquals(options.password, "password");
-});
-
-suite.test("connect with URL", async () => {
-  const options = parseURL("redis://127.0.0.1:7003/1");
-  const tempClient = await connect(options);
-  assertEquals(tempClient.isConnected, true);
-  assertEquals(tempClient.isClosed, false);
-  assertEquals(await tempClient.quit(), "OK");
-  assertEquals(tempClient.isConnected, false);
-  assertEquals(tempClient.isClosed, true);
 });
 
 suite.runTests();
